Extracted and moved some logic from AuditHelper to TransactionManager
